[BUG: #66] include/ppsi/jiffies.h: Avoid undefined behavior from signed overflow in time_after
According to the C standard 3.4.3p3, overflow of a signed integer results
in undefined behavior.  This commit therefore changes the definitions
of time_after(), time_after_eq() to avoid this undefined behavior.
The trick is that the subtraction is done using unsigned arithmetic,
which according to 6.2.5p9 cannot overflow because it is defined as modulo
arithmetic.

Note that the C standard considers the cast from unsigned to
signed to be implementation-defined, see 6.3.1.3p3.

For more information see:
https://github.com/torvalds/linux/commit/5a581b367b5df0531265311fc681c2abd377e5e6

The problem was seen with GCC 11.x for RISC-V.

+++ include/ppsi/jiffies.h
@@ -29,13 +29,13 @@
 #define time_after(a,b)		\
 	(typecheck(unsigned long, a) && \
 	 typecheck(unsigned long, b) && \
-	 ((long)(b) - (long)(a) < 0))
+	 ((long)((b) - (a)) < 0))
 #define time_before(a,b)	time_after(b,a)
 
 #define time_after_eq(a,b)	\
 	(typecheck(unsigned long, a) && \
 	 typecheck(unsigned long, b) && \
-	 ((long)(a) - (long)(b) >= 0))
+	 ((long)((a) - (b)) >= 0))
 #define time_before_eq(a,b)	time_after_eq(b,a)
 #endif /* ifndef */
 #endif
